feat(Database): 데이터베이스 관련 로직 삭제
