+++ src/main/java/janggi/database/dao/TeamDao.java
@@ -1,4 +1,4 @@
-package janggi.dao;
+package janggi.database.dao;
 
 import janggi.domain.Team;
 import java.sql.Connection;
